+++ changelog.md
@@ -3,7 +3,8 @@ v1.2 (2026-xx-xx):
 * Fix potential desync issue w/ team revive queue
 * Fix team lives persisting across sessions
 * Initialize properly if loaded mid-level
-* Improve coop HUD issues, sorting by distance w/ default 4p max
+* Improve coop HUD, sorting by distance and combining bot groups
+* Default `cd_hudmaxplayers` to 4, instead of 8
 * Add `cd_hudsorttime` for HUD sort interval
 * Add `pinplayer2` / `unpinplayer2` for splitscreen use
 * Allow partial names for `pinplayer` etc. instead of only numbers