+++ changelog.md
@@ -11,6 +11,7 @@ v1.2 (2026-xx-xx):
 * Allow partial names for `pinplayer` etc. instead of only numbers
 * Fix getting infinitely juggled by tagged enemies when super or invincible (whoops, lol)
 * Award % of Enemy Goal after warping back to start at least once
+* Fix revive mechanic not clearing player `outofcoop` status
 * Code cleanup!
 
 v1.1 (2021-06-11):